expand volume accuracy from 8bit to 16bit; allow channel volume values to be > 1, and clamp the resulting volume
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@10482 d7cf8633-e32d-0410-b094-e92efae38249

+++ snd_main.h
@@ -84,8 +84,8 @@ struct sfx_s
 
 typedef struct channel_s
 {
-	short			listener_volume [SND_LISTENERS];	// 0-255 volume per speaker
-	int				master_vol;		// 0-255 master volume
+	int			listener_volume [SND_LISTENERS];	// 0-65536 volume per speaker
+	int				master_vol;		// 0-65536 master volume
 	sfx_t			*sfx;			// sfx number
 	unsigned int	flags;			// cf CHANNELFLAG_* defines
 	int				pos;			// sample position in sfx, negative values delay the start of the sound playback
